move RangeFactory to package range
