+++ src/main/java/org/repozoo/commons/range/factory/impl/LocalDateRange.java
@@ -1,7 +1,7 @@
 package org.repozoo.commons.range.factory.impl;
 
 import org.repozoo.commons.range.Range;
-import org.repozoo.commons.range.factory.RangeFactory;
+import org.repozoo.commons.range.RangeFactory;
 
 import java.time.LocalDate;
 import java.util.function.UnaryOperator;
