+++ src/test/java/org/repozoo/commons/range/impl/ExamplesTest.java
@@ -1,7 +1,7 @@
 package org.repozoo.commons.range.impl;
 
 import org.repozoo.commons.range.Range;
-import org.repozoo.commons.range.factory.RangeFactory;
+import org.repozoo.commons.range.RangeFactory;
 import org.repozoo.commons.range.factory.impl.LocalDateRange;
 import org.repozoo.commons.range.RangeSet;
 import org.junit.jupiter.api.Test;
